+++ tests/View/RelativePath/views-parent/template/with-extension.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * @var \Yiisoft\View\View $this
+ */
+
+echo 'THIS IS TEMPLATE. ';
+
+echo $this->render('../_header.php');